+++ src/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './file-function';
 export * from './getFilesFromGetAsEntry';
+export * from './getStructureFiles';